Adds a way to delete pipelines but leave their output repos around

+++ src/server/pps/server/api_server.go
@@ -2552,9 +2552,22 @@ func (a *apiServer) deletePipeline(pachClient *client.APIClient, request *pps.De
 		if err := a.authorizePipelineOp(pachClient, pipelineOpDelete, pipelineInfo.Input, pipelineInfo.Pipeline.Name); err != nil {
 			return nil, err
 		}
-		// delete the pipeline's output repo
-		if err := pachClient.DeleteRepo(request.Pipeline.Name, request.Force); err != nil {
-			return nil, err
+		if request.KeepRepo {
+			// Remove branch provenance (pass branch twice so that it continues to point
+			// at the same commit, but also pass empty provenance slice)
+			if err := pachClient.CreateBranch(
+				request.Pipeline.Name,
+				pipelineInfo.OutputBranch,
+				pipelineInfo.OutputBranch,
+				nil,
+			); err != nil {
+				return nil, err
+			}
+		} else {
+			// delete the pipeline's output repo
+			if err := pachClient.DeleteRepo(request.Pipeline.Name, request.Force); err != nil {
+				return nil, err
+			}
 		}
 	}
 
@@ -2622,13 +2635,15 @@ func (a *apiServer) deletePipeline(pachClient *client.APIClient, request *pps.De
 		return nil
 	})
 	// Delete cron input repos
-	pps.VisitInput(pipelineInfo.Input, func(input *pps.Input) {
-		if input.Cron != nil {
-			eg.Go(func() error {
-				return pachClient.DeleteRepo(input.Cron.Repo, request.Force)
-			})
-		}
-	})
+	if !request.KeepRepo {
+		pps.VisitInput(pipelineInfo.Input, func(input *pps.Input) {
+			if input.Cron != nil {
+				eg.Go(func() error {
+					return pachClient.DeleteRepo(input.Cron.Repo, request.Force)
+				})
+			}
+		})
+	}
 	if err := eg.Wait(); err != nil {
 		return nil, err
 	}
